# apps/example/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_TaskRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[TaskRead]",
        "type": "object"
      },
...
      "TaskCreate": {
        "properties": {
          "status": {
            "default": "open",
            "maxLength": 20,
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TaskCreate",
        "type": "object"
      },
      "TaskRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "modified_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified By Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "version",
          "created_at",
          "updated_at",
          "created_by_id",
          "modified_by_id"
        ],
        "title": "TaskRead",
        "type": "object"
      },
...
      "TicketRequest": {
        "properties": {
          "channel": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Channel",
            "type": "string"
          },
          "transport": {
            "enum": [
              "sse",
              "websocket"
            ],
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "transport"
        ],
        "title": "TicketRequest",
        "type": "object"
      },
      "TicketResponse": {
        "properties": {
          "channel": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Channel",
            "type": "string"
          },
          "expires_in": {
            "exclusiveMinimum": 0.0,
            "title": "Expires In",
            "type": "integer"
          },
          "ticket": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Ticket",
            "type": "string"
          },
          "transport": {
            "enum": [
              "sse",
              "websocket"
            ],
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "expires_in",
          "channel",
          "transport"
        ],
        "title": "TicketResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Terp example app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/realtime/sse/{channel_name}": {
      "get": {
        "operationId": "realtime.subscribe_sse",
        "parameters": [
          {
            "in": "path",
            "name": "channel_name",
            "required": true,
            "schema": {
              "title": "Channel Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ticket",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Ticket",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start streaming live updates for one channel",
        "tags": [
          "realtime"
        ]
      }
    },
    "/api/v1/realtime/tickets": {
      "post": {
        "operationId": "realtime.mint_ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a short-lived pass that lets someone open a live connection",
        "tags": [
          "realtime"
        ]
      }
    },
    "/api/v1/tasks/": {
      "get": {
        "operationId": "tasks.list_tasks",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Maximum rows to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum rows to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List every task",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "tasks.create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a new task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}